The DataType

The main data-type

Core.AbstractFloat — Method

AbstractFloat(x::ℤd{T,d}) where {T <: Integer, d}

Converts a given ℤd{T, d} to a general Float. Used mostly internal by Julia. See convert(::Type{K}, x::ℤd{T, d}) for more details on conversion.

Core.Int — Method

Int(x::ℤd{T,d}) where {T <: Integer, d}

Converts a given ℤd{T, d} to a general Integer. Used mostly internal by Julia. See convert(::Type{K}, x::ℤd{T, d}) for more details on conversion.

Zd_Arithmetics.ℤd — Type

ℤd{T <: Integer, d} <: Integer

The following immutable struct stores a value of type T <: Integer modulo d.

The type T is customizable, but should be choosen as the smallest Integer type that fits d. For normal use cases, Int8 is enough.

Example:

ℤd{Int8, 2}(1) # stores a 1
ℤd{Int8, 2}(2) # stores mod(2,2) = 0

For convenience it is recommended to use type aliases if you (mostly) work with fixed d:

# Define a type alias for the integers modulo two
const ℤ₂ = ℤd{Bool, 2}
# Define a type alias for the integers modulo three
const ℤ₃ = ℤd{Int8, 3}

Another method is to use the provided type aliases via import:

ℤ₂ = Zd_Arithmetics.Z2
ℤ₂ = Zd_Arithmetics.ℤ₂

or

import Zd_Arithmetics.Z2
import Zd_Arithmetics.ℤ₂

Zd_Arithmetics.ℤd — Method

ℤd{T, d}(x::ℤd{T, d′}; force=false)::ℤd{T,d} where {T <: Integer, d, d′}

Upwards converting from ℤd′ to ℤd for d′ < d (possible, because $\mathbb{Z}_{d'} \subset \mathbb{Z}_d$ for $d' < d$).

For d == d′ the original x will be returned.

For d > d′ an error is thrown if the optional parameter force is not true (default: false). If the force flag is enabled, then the value of x will be mapped to $x \mod d$ regardless of d or d′ (possible loss of information).

Non-exported aliases

To use one of the following type aliases, just import/declare them like this

Z2 = Zd_Arithmetics.Z2 # allows renaming
# or
import Zd_Arithmetics.Z2

Apart from that, you can always define your own alias as described above. For the $\mathbb{Z}_2$ case it's best to import the type alias from the package, since then the specialized simplified operations can be used.

Zd_Arithmetics.ℤ₂ — Type

ℤ₂ = ℤd{Bool, 2}
Z2 = ℤ₂

Type aliases for integers modulo 2.

The type T is set to Bool instead of Int8 for performance-reasons. The simplified operations for the ℤ₂ case can be found at the bottom of operations.jl.

Zd_Arithmetics.ℤ₃ — Type

ℤ₃ = ℤd{Int8, 3}
Z3 = ℤ₃

Type aliases for integers modulo 3.

Zd_Arithmetics.ℤ₄ — Type

ℤ₄ = ℤd{Int8, 4}
Z4 = ℤ₄

Type aliases for integers modulo 4.

Zd_Arithmetics.ℤ₅ — Type

ℤ₅ = ℤd{Int8, 5}
Z5 = ℤ₅

Type aliases for integers modulo 5.

Zd_Arithmetics.ℤ₆ — Type

ℤ₆ = ℤd{Int8, 6}
Z6 = ℤ₆

Type aliases for integers modulo 6.