Operations

Base.:* — Method

+(x::ℤ₂, y::ℤ₂)

Special case of multiplication in ℤ₂ using Booleans.

Base.:* — Method

*(a::Integer, b::ℤd{T, d}) where {T <: Integer, d}

Enables ℤ × ℤd → ℤd multiplication.

Base.:* — Method

*(a::ℤd{T, d}, b::ℤd{T, d}) where {T <: Integer, d}

Multiplies the values of a and b modulo d and returns a new ℤd{T, d}

Base.:+ — Method

+(x::ℤ₂, y::ℤ₂)

Special case of addition in ℤ₂ using Booleans.

Base.:+ — Method

+(a::ℤd{T, d}, b::ℤd{T, d}) where {T <: Integer, d}

Adds the values of a and b modulo d and returns a new ℤd{T, d}

Base.:- — Method

-(x::ℤ₂, y::ℤ₂)

Special case of subtraction in ℤ₂ using Booleans.

Base.:- — Method

-(a::ℤd{T, d}, b::ℤd{T, d}) where {T <: Integer, d}

Subtracts the value of b from a modulo d and returns a new ℤd{T, d} (always positive).

Base.:- — Method

-(x::ℤd{T,d}) where {T <: Integer, d}

Negating x in ℤd yields again x.

Base.convert — Method

convert(::Type{K}, x::ℤd{T, d}) where {K <: Number, T <: Integer, d}

Converts a given ℤd{T, d} to a general number without mod d-logic.

This is usefull for many more complex operations and definitions of functions that live outside ℤd. Also, some external functions are not reimplemented, therefore they can only be used if ℤd{T, d} is treated like a normal number. If the function-result should again map to ℤd{T, d}, remember to convert the result back to ℤd{T, d}.

Info

If the external function returns floating point values, it's necessary to convert them back to Integers first using something like

# rounding a float x to the next integer and converting it back to an integer.
Int(round(x; digits=0))

Base.inv — Method

inv(A::Matrix{ℤd{T, d}})

Invert a matrix A in ℤd by multiplying the inverse in ℝ with the determinant.

Base.promote_rule — Method

Promote the ℤ₂ type over Bool.

Base.promote_rule — Method

Promote the ℤd{T, d} type over T.

Base.show — Method

show(io::IO, x::ℤ₂)

Special case of pretty-printing in ℤ₂ using Booleans.

Base.show — Method

show(io::IO, x::ℤd{T, d}) where {T <: Integer, d}

Pretty-printing for ℤd as Integers.

Base.trailing_zeros — Method

trailing_zeros(x::ℤd{T, d}) where {T <: Integer, d}

Internal Julia function needed in some cases, like when using Sets. Passes the call down to the value of x.